create passkey base resolver
