+++ src/packages/auth/src/authentication/entities/authentication.ts
@@ -1,6 +1,7 @@
 export interface AuthenticationBaseEntity<T> {
 	id: string;
 	type: string;
+	userId: string;
 	data: T;
 	createdAt: Date;
 }
