+++ src/packages/auth/src/authentication/providers/passkey/index.ts
@@ -1,1 +1,2 @@
+export * from './base-resolver';
 export * from './resolver';
